# opt/idxconstraint: optimize computed column constraint derivation