#### opt/bench: add benchmark like "internal-upsert-txn-stats" statement

A new opt benchmark has been added that mimics the
"internal-upsert-txn-stats" query executed by the `persistedsqlstats`
package.

Release note: None

#### opt/idxconstraint: correctly reset skip flag

The `skipComputedColPredDerivation` flag is now correctly reset to its
previous value with `defer`. This caused no known correctness or
performance issues—fixing it guards against future unexpected behavior.

Release note: None

#### opt/idxconstraint: reorder computed column derivation conditions

Some `if` conditionals have been reordered from least expensive to most
expensive to benefit from short-circuiting.

Release note: None

#### opt/idxconstraint: optimize computed column constraint derivation

Building index constraints on computed columns is currently very
expensive. Prior to this commit, this derivation occurred roughly
`C * E` times for every index, where `C` is the number of key columns in
the index and `E` is the number of filter expressions. For each suffix
of key columns, a computed column constraint was attempted to be
derived.

This commit reduces this by only trying to derive computed column
constraints on key column suffixes that actually contain computed
columns. In many cases, such as hash-sharded indexes, this effectively
reduces computed column constraint derivation to be attempted `E` times
per index, because in a hash-sharded index the first column is commonly
the only computed column.

Informs #134587

Release note: None