@mgartner 2024-11-08
The newly added benchmarks show the improvement:

```
name                                             old time/op    new time/op    delta
Phases/comp-pk/Simple/Explore                       737µs ± 1%     561µs ±17%  -23.79%  (p=0.008 n=5+5)
Phases/comp-pk/Prepared/Explore                     585µs ± 0%     381µs ± 0%  -34.85%  (p=0.008 n=5+5)
Phases/comp-insert-on-conflict/Simple/Explore       976µs ± 1%     770µs ± 1%  -21.09%  (p=0.008 n=5+5)
Phases/comp-insert-on-conflict/Prepared/Explore     655µs ± 0%     447µs ± 0%  -31.79%  (p=0.016 n=5+4)

name                                             old alloc/op   new alloc/op   delta
Phases/comp-pk/Simple/Explore                       400kB ± 0%     271kB ± 0%  -32.25%  (p=0.008 n=5+5)
Phases/comp-pk/Prepared/Explore                     335kB ± 0%     206kB ± 0%  -38.49%  (p=0.008 n=5+5)
Phases/comp-insert-on-conflict/Simple/Explore       594kB ± 0%     463kB ± 0%  -22.16%  (p=0.008 n=5+5)
Phases/comp-insert-on-conflict/Prepared/Explore     374kB ± 0%     242kB ± 0%  -35.22%  (p=0.008 n=5+5)

name                                             old allocs/op  new allocs/op  delta
Phases/comp-pk/Simple/Explore                       3.80k ± 0%     2.28k ± 0%  -40.06%  (p=0.008 n=5+5)
Phases/comp-pk/Prepared/Explore                     3.48k ± 0%     1.96k ± 0%  -43.77%  (p=0.008 n=5+5)
Phases/comp-insert-on-conflict/Simple/Explore       4.53k ± 0%     2.99k ± 0%  -33.98%  (p=0.008 n=5+5)
Phases/comp-insert-on-conflict/Prepared/Explore     3.69k ± 0%     2.15k ± 0%  -41.73%  (p=0.000 n=5+4)
```